+++ mimesis/builtins/de.py
@@ -14,7 +14,7 @@ class GermanySpecProvider(BaseSpecProvider):
     def __init__(self, seed: Seed = None):
         """Initialize attributes."""
         super().__init__(locale='de', seed=seed)
-        self.pull(self._datafile)
+        self._pull(self._datafile)
 
     class Meta:
         """The name of the provider."""