+++ mimesis/builtins/it.py
@@ -18,7 +18,7 @@ class ItalySpecProvider(BaseSpecProvider):
     def __init__(self, seed: Seed = None):
         """Initialize attributes."""
         super().__init__(locale='it', seed=seed)
-        self.pull(self._datafile)
+        self._pull(self._datafile)
 
     class Meta:
         """The name of the provider."""