+++ mimesis/providers/science.py
@@ -20,7 +20,7 @@ class Science(BaseDataProvider):
         """
         super().__init__(*args, **kwargs)
         self._datafile = 'science.json'
-        self.pull(self._datafile)
+        self._pull(self._datafile)
 
     class Meta:
         """Class for metadata."""