+++ tests/test_providers/test_base.py
@@ -55,14 +55,14 @@ class TestBase(object):
     )
     def test_pull(self, locale, city):
         data_provider = BaseDataProvider(locale)
-        data_provider.pull('address.json')
+        data_provider._pull('address.json')
         assert city in data_provider._data['city']
 
     @pytest.mark.parametrize('locale', LIST_OF_LOCALES)
     def test_pull_raises(self, locale):
         data_provider = BaseDataProvider(locale=locale)
         with pytest.raises(FileNotFoundError):
-            data_provider.pull('something.json')
+            data_provider._pull('something.json')
 
     def test_update_dict(self, base_data_provider):
         first = {